Addressed Stylecop comment warnings.
