+++ solutions/Virtual-Assistant/src/csharp/skills/pointofinterestskill/pointofinterestskill/Models/Foursquare/Hours.cs
@@ -1,10 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace PointOfInterestSkill.Models.Foursquare
 {
@@ -24,14 +20,5 @@ namespace PointOfInterestSkill.Models.Foursquare
 
         [JsonProperty(PropertyName = "dayData")]
         public object[] DayData { get; set; }
-    }
-
-    public class Richstatus
-    {
-        [JsonProperty(PropertyName = "entities")]
-        public object[] Entities { get; set; }
-
-        [JsonProperty(PropertyName = "text")]
-        public string Text { get; set; }
     }
 }
