+++ solutions/Virtual-Assistant/src/csharp/skills/pointofinterestskill/pointofinterestskill/Models/Foursquare/Menu.cs
@@ -1,10 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace PointOfInterestSkill.Models.Foursquare
 {
@@ -18,7 +14,8 @@ namespace PointOfInterestSkill.Models.Foursquare
 
         [JsonProperty(PropertyName = "anchor")]
         public string Anchor { get; set; }
-                [JsonProperty(PropertyName = "url")]
+
+        [JsonProperty(PropertyName = "url")]
         public string Url { get; set; }
 
         [JsonProperty(PropertyName = "mobileUrl")]
