+++ solutions/Virtual-Assistant/src/csharp/skills/pointofinterestskill/pointofinterestskill/Models/Foursquare/Source.cs
@@ -1,8 +1,9 @@
-﻿using Newtonsoft.Json;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+using Newtonsoft.Json;
 
 namespace PointOfInterestSkill.Models.Foursquare
 {
-
     public class Source
     {
         [JsonProperty(PropertyName = "name")]
