+++ solutions/Virtual-Assistant/src/csharp/skills/pointofinterestskill/pointofinterestskill/Models/Foursquare/Venuepage.cs
@@ -1,8 +1,6 @@
-﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+using Newtonsoft.Json;
 
 namespace PointOfInterestSkill.Models.Foursquare
 {
