+++ solutions/Virtual-Assistant/src/csharp/skills/pointofinterestskill/pointofinterestskill/Models/PointofInterestModel.cs
@@ -114,93 +114,140 @@ namespace PointOfInterestSkill.Models
         /// Gets or sets the id of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The id of this point of interest.
+        /// </value>
         public string Id { get; set; }
 
         /// <summary>
         /// Gets or sets the thumbnail image url of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The image URL of this point of interest.
+        /// </value>
         public string ImageUrl { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The name of this point of interest.
+        /// </value>
         public string Name { get; set; }
 
         /// <summary>
         /// Gets or sets the address of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The city of this point of interest.
+        /// </value>
         public string City { get; set; }
 
         /// <summary>
         /// Gets or sets the address line of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The street address of this point of interest.
+        /// </value>
         public string Street { get; set; }
 
         /// <summary>
         /// Gets or sets the geolocation of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The geolocation of this point of interest.
+        /// </value>
         public LatLng Geolocation { get; set; }
 
         /// <summary>
         /// Gets or sets the estimated time of arrival to the point of interest.
         /// Availability: Azure Maps.
         /// </summary>
+        /// <value>
+        /// The ETA to this point of interest.
+        /// </value>
         public string EstimatedTimeOfArrival { get; set; }
 
         /// <summary>
         /// Gets or sets the distance to the point of interest.
         /// Availability: Azure Maps.
         /// </summary>
+        /// <value>
+        /// The distance to this point of interest.
         public string Distance { get; set; }
 
         /// <summary>
         /// Gets or sets the rating of the point of interest.
         /// Availability: Foursquare.
         /// </summary>
+        /// <value>
+        /// The rating of this point of interest.
+        /// </value>
         public string Rating { get; set; }
 
         /// <summary>
         /// Gets or sets the number of ratings of the point of interest.
         /// </summary>
+        /// <value>
+        /// The rating count of this point of interest.
+        /// </value>
         public int RatingCount { get; set; }
 
         /// <summary>
         /// Gets or sets the price level of the point of interest.
         /// Availability: Foursquare.
         /// </summary>
+        /// <value>
+        /// The price of this point of interest.
+        /// </value>
         public int Price { get; set; }
 
         /// <summary>
         /// Gets or sets the hours of the point of interest.
         /// Availability: Foursquare.
         /// </summary>
+        /// <value>
+        /// The open hours of this point of interest.
+        /// </value>
         public string Hours { get; set; }
 
         /// <summary>
-        /// Gets or sets the top category ofthe point of interest.
+        /// Gets or sets the top category of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The top category of this point of interest.
+        /// </value>
         public string Category { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the point of interest.
         /// Availability: Azure Maps, Foursquare.
         /// </summary>
+        /// <value>
+        /// The provider of this point of interest.
+        /// </value>
         public string Provider { get; set; }
 
         /// <summary>
-        /// Gets or sets the option number of the point of interest.
+        /// Gets or sets the index number of the point of interest.
         /// </summary>
+        /// <value>
+        /// The index of this point of interest.
+        /// </value>
         public int Index { get; set; }
 
         /// <summary>
-        /// Gets or sets the formatted string for available details.
+        /// Gets the formatted string for available details.
         /// </summary>
+        /// <value>
+        /// The available details formatted string of this point of interest.
+        /// </value>
         public string AvailableDetails
         {
             get
